feat: reimplement pegasus_value_generator
