+++ src/base/value_schema_v0.cpp
@@ -19,6 +19,8 @@
 
 #include "value_schema_v0.h"
 
+#include <dsn/utility/endians.h>
+#include <dsn/c/api_utilities.h>
 #include <dsn/dist/fmt_logging.h>
 #include <dsn/utility/smart_pointers.h>
